seeded data and made api call
