+++ server/api/user/user.model.js
@@ -16,16 +16,7 @@ var UserSchema = new Schema({
   hashedPassword: String,
   email: {type: String, lowercase: true },
   salt: String,
-  shows: [{
-  	show: {type: Schema.Types.ObjectId},
-  	previousEpisodeS: {type: Number},
-  	previousEpisodeN: {type: Number},
-  	episodesWatched: {type: Number},
-  	favorites: [{
-  		episodeS: {type: [Number]},
-  		episodeN: {type: [Number]}
-  	}]
-  }]
+  shows: {}
 });
 
 /**
