+++ server/config/apiKey.js
@@ -3,5 +3,5 @@
 // Use apiKey.js for apiKeys used
 
 module.exports = {
-  tvdb: 'haha'
+  tvdb: '49D3CC1F3C6665A8'
 };
